Merged PR 15806: Added correlation id provider concept to permit delayed-evaluation of
Added correlation id provider concept to permit delayed-evaluation of correlation id.  As the ASP.NET pipeline instantiates all types in the pipeline before the pipeline is executed, this permits middleware to read the correlation id from the request headers and a later ICorrelationIdProvider to hand that to the IMessagePublisher.  Marked the string constructor overload as obsolete to later permit simpler IoC setup for structuremap especially.